swap implementation and shim to ease potential deprecation
